requirements.txt
python-telegram-bot[job-queue]>=21.0,<22
python-dotenv==1.0.1
httpx>=0.26.0
pytest>=7.0
pytest-asyncio>=0.21
